refactor: split parallel coordinators into dedicated package
