+++ projects/04-llm-adapter/adapter/core/parallel/coordinators/__init__.py
@@ -1,0 +1,18 @@
+"""Parallel coordinator implementations."""
+
+from __future__ import annotations
+
+from ...parallel_state import ProviderFailureSummary, build_cancelled_result
+from .all import _ParallelAllCoordinator
+from .any import _ParallelAnyCoordinator
+from .base import _ParallelCoordinatorBase, _is_parallel_any_mode, _normalize_mode_value
+
+__all__ = [
+    "ProviderFailureSummary",
+    "_ParallelAllCoordinator",
+    "_ParallelAnyCoordinator",
+    "_ParallelCoordinatorBase",
+    "_is_parallel_any_mode",
+    "_normalize_mode_value",
+    "build_cancelled_result",
+]
